Revert "updated edit by hand, to much changes for sourcetree"
This reverts commit 889319466ef2e5ccf62d8ea767fd24da469ef3fb.

+++ contrib/scintilla/cocoa/QuartzTextStyle.h
@@ -61,17 +61,17 @@ public:
 
 	float getAscent() const
 	{
-		return static_cast<float>(::CTFontGetAscent(fontRef));
+		return ::CTFontGetAscent(fontRef);
 	}
 
 	float getDescent() const
 	{
-		return static_cast<float>(::CTFontGetDescent(fontRef));
+		return ::CTFontGetDescent(fontRef);
 	}
 
 	float getLeading() const
 	{
-		return static_cast<float>(::CTFontGetLeading(fontRef));
+		return ::CTFontGetLeading(fontRef);
 	}
 
 	void setFontRef(CTFontRef inRef, int characterSet_)
